pizza: 15: Add Domain and Range to hasBase
WARNING: Although we have specified the domains and ranges of various
properties for the purposes of this tutorial, we generally advise
against doing this. The fact that domain and range conditions do not
behave as constraints and the fact that they can cause ‘unexpected’
classification results can lead problems and unexpected side effects.
These problems and side effects can be particularly difficult to track
down in a large ontology.

Signed-off-by: Hernán Gonzalez <hernan@vanguardiasur.com.ar>