fhkb: 21: Siblings option two
Option one is obviously the least effort; we only have to assert the
same parentage facts as we already have; then the sub-property chains do
the rest. It works OK for hasSibling, but we cannot do brothers and
sisters adequately; we need Man and hasSibling > isBrotherOf and we
cannot do that implication. This means we cannot ask the questions we
need to ask.

So, we do option two, even though it is hard work and is still not
perfect for query answering, even though we have gone for a sparse
assertion mode. Doing full sibling assertion would work, but is a lot of
effort.

Signed-off-by: Hernán Gonzalez <hernan@vanguardiasur.com.ar>